Move is_dead to common::mod.rs
